refactor: extract Google People API client to shared module
