+++ backend/src/services/export/exports/google/index.ts
@@ -1,9 +1,11 @@
 import { GaxiosError } from 'gaxios';
-import { google, people_v1 } from 'googleapis';
 import ENV from '../../../../config';
 import { ContactFrontend } from '../../../../db/types';
 import logger from '../../../../utils/logger';
-import googleOAuth2Client from '../../../OAuth2/google';
+import {
+  getPeopleService,
+  warmupSearchIndex
+} from '../../../OAuth2/googlePeopleClient';
 import {
   ExportStrategy,
   ExportType,
@@ -12,67 +14,10 @@ import {
 } from '../../types';
 import GoogleContactsSession from './contacts-api';
 
-export default class GoogleContactsExport
-  implements ExportStrategy<ContactFrontend>
-{
+export default class GoogleContactsExport implements ExportStrategy<ContactFrontend> {
   readonly type = ExportType.GOOGLE_CONTACTS;
 
   readonly contentType = 'application/json';
-
-  private static async getPeopleService(
-    options: ExportOptions['googleContactsOptions']
-  ) {
-    const accessToken = options?.accessToken;
-    const refreshToken = options?.refreshToken;
-
-    if (!accessToken && !refreshToken) {
-      throw new Error('Invalid credentials.');
-    }
-
-    const oauth2Client = new google.auth.OAuth2(
-      ENV.GOOGLE_CLIENT_ID,
-      ENV.GOOGLE_SECRET
-    );
-
-    const token = googleOAuth2Client.createToken({
-      access_token: accessToken,
-      refresh_token: refreshToken
-    });
-
-    if (token.expired(1000) && !refreshToken) {
-      throw new Error('Invalid credentials.');
-    }
-
-    oauth2Client.setCredentials({
-      access_token: accessToken,
-      refresh_token: refreshToken
-    });
-
-    const peopleService = google.people({
-      version: 'v1',
-      auth: oauth2Client
-    });
-
-    // Update cache and validate credentials
-    await GoogleContactsExport.warmupSearchIndex(peopleService);
-
-    return peopleService;
-  }
-
-  private static async warmupSearchIndex(
-    service: people_v1.People
-  ): Promise<void> {
-    await service.people.searchContacts({
-      query: '',
-      readMask: 'names,emailAddresses'
-    });
-
-    await service.people.connections.list({
-      resourceName: 'people/me',
-      pageSize: 1,
-      personFields: 'metadata'
-    });
-  }
 
   async export(
     contacts: ContactFrontend[],
@@ -83,7 +28,11 @@ export default class GoogleContactsExport
 
     try {
       const opts = options.googleContactsOptions;
-      const service = await GoogleContactsExport.getPeopleService(opts);
+      const service = await getPeopleService({
+        accessToken: opts.accessToken,
+        refreshToken: opts.refreshToken
+      });
+      await warmupSearchIndex(service);
       const session = new GoogleContactsSession(
         service,
         ENV.APP_NAME,
